sys/psa_crypto: Reintroduce HMAC hardware acceleration
The nrf52 supports hardware acceleration of HMAC algorithms,
so reintroduce support for periph drivers to provide a special
hardware-accelerated HMAC backend.

Please note that the multi-step PSA MAC API is currently not
supported by the periph backend, so programs which depend on
this API need to select the generic HMAC backend.

Signed-off-by: Armin Wolf <W_Armin@gmx.de>